fix: Error With Invalid IDs
- Fixed an error which occurred with most requests where if an invalid ID was sent, it would not send a response back
- Updated template data to better match the functionality of asset data
- Reworked asset category updating to make it more reliable and consistent
- Erased a lot of unnecessary commented sections of code

+++ app/controllers/alertType.controller.js
@@ -47,6 +47,9 @@ exports.findAll = (req, res) => {
 // Find a single AlertType with an id
 exports.findOne = (req, res) => {
   const id = req.params.id;
+  if (isNaN(parseInt(id))) return res.status(400).send({
+    message: "Invalid alert type id!",
+  });
 
   AlertType.findByPk(id)
   .then((data) => {
@@ -68,6 +71,9 @@ exports.findOne = (req, res) => {
 // Update an AlertType by the id in the request
 exports.update = (req, res) => {
   const id = req.params.id;
+  if (isNaN(parseInt(id))) return res.status(400).send({
+    message: "Invalid alert type id!",
+  });
 
   AlertType.update(req.body, { where: { id } })
   .then((num) => {
@@ -91,6 +97,9 @@ exports.update = (req, res) => {
 // Delete an AlertType with the specified id in the request
 exports.delete = (req, res) => {
   const id = req.params.id;
+  if (isNaN(parseInt(id))) return res.status(400).send({
+    message: "Invalid alert type id!",
+  });
 
   AlertType.destroy({ where: { id } })
   .then((num) => {
@@ -110,20 +119,3 @@ exports.delete = (req, res) => {
     });
   });
 };
-
-// Delete all AlertTypes from the database.
-// exports.deleteAll = (req, res) => {
-//   AlertType.destroy({
-//     where: {},
-//     truncate: false,
-//   })
-//   .then((nums) => {
-//     res.send({ message: `${nums} alert types were deleted successfully!` });
-//   })
-//   .catch((err) => {
-//     res.status(500).send({
-//       message:
-//         err.message || "Some error occurred while removing all alert types.",
-//     });
-//   });
-// };
